+++ modules/db/src/main/groovy/com/intellisrc/db/PoolConnector.groovy
@@ -4,6 +4,8 @@ import com.intellisrc.core.Log
 import com.intellisrc.db.jdbc.JDBC
 import groovy.transform.CompileStatic
 
+import java.sql.Connection
+
 @CompileStatic
 /**
  * Implements a Connector to be used in DBPool
@@ -44,12 +46,12 @@ class PoolConnector implements Connector {
 
 	@Override
 	List<String> getTables() {
-		return currentConnector?.tables
+		return open() ? currentConnector.tables : []
 	}
 
 	@Override
 	List<ColumnInfo> getColumns(String table) {
-		return open() ? currentConnector?.getColumns(table) : []
+		return open() ? currentConnector.getColumns(table) : []
 	}
 
 	@Override
@@ -73,25 +75,38 @@ class PoolConnector implements Connector {
 	}
 
 	@Override
+	void clear(Connection connection) {
+		if(open) {
+			currentConnector.clear(connection)
+		}
+	}
+
+	@Override
 	boolean close() {
-		pool?.returnConnectionToPool(currentConnector)
-		Log.v( "DB returned.")
-		return true
+		boolean returned = false
+		if(currentConnector) {
+			pool?.returnConnectionToPool(currentConnector)
+			Log.v("DB returned.")
+			returned = true
+		}
+		return returned
 	}
 
 	@Override
 	ResultStatement execute(Query query, boolean silent) {
-		return currentConnector?.execute(query, silent)
+		return open() ? currentConnector.execute(query, silent) : null
 	}
 
     @Override
     boolean commit(List<Query> queries) {
-        return currentConnector?.commit(queries)
+        return open() ? currentConnector.commit(queries) : false
     }
 
 	@Override
 	void rollback() {
-		currentConnector?.rollback()
+		if(open()) {
+			currentConnector.rollback()
+		}
 	}
 
 	@Override