+++ modules/db/src/main/groovy/com/intellisrc/db/jdbc/SQLServer.groovy
@@ -1,7 +1,6 @@
 package com.intellisrc.db.jdbc
 
 import com.intellisrc.core.Config
-import com.intellisrc.db.auto.AutoJDBC
 import groovy.transform.CompileStatic
 
 /**
@@ -21,15 +20,19 @@ class SQLServer extends JDBCServer {
     String driver = "com.microsoft.sqlserver.jdbc.SQLServerDriver"
     // SQLServer specific parameters:
     // Most common:
-    boolean ssl = false
     boolean useWinLogin = false
     boolean trustCert = false
+
+    //encrypt : true // for drivers 10.2 and below
+    //encrypt : "strict" // For drivers 11.2 or above
+    boolean secure = false
+    boolean strict = false // Only if secure = true
     // https://docs.microsoft.com/en-us/sql/connect/jdbc/setting-the-connection-properties?view=sql-server-ver15
     // You may add more parameters as needed (values shown below are default values)
     @Override
     Map getParameters() {
         return Config.get("db.sqlserver.params", [
-            encrypt : ssl,
+            encrypt : secure && strict ? "strict" : secure.toString(),
             integratedSecurity : useWinLogin,
             loginTimeout : 15,
             trustServerCertificate : trustCert,