+++ modules/db/src/test/groovy/com/intellisrc/db/jdbc/JDBCTest.groovy
@@ -10,7 +10,8 @@ import com.intellisrc.term.TableMaker
 import spock.lang.IgnoreIf
 import spock.lang.Specification
 
-import static com.intellisrc.db.Query.SortOrder.*
+import static com.intellisrc.db.Query.SortOrder.ASC
+import static com.intellisrc.db.Query.SortOrder.DESC
 
 /**
  * @since 2022/01/20.
@@ -56,6 +57,7 @@ abstract class JDBCTest extends Specification {
     def setup() {
         DB db = getDB().connect()
         db.dropAllTables()
+        db.clearCache()
         db.close()
     }
 
@@ -217,10 +219,13 @@ abstract class JDBCTest extends Specification {
             ],[8,9])
         then: "Update with map criteria"
             assert db.table(table).update([ active: false ], [ active: true ])
-        then: "Update with id, multiple"
+        then: "Update multiple"
             assert db.table(table).update([
-                3 : [ version : 3.4f ],
-                4 : [ version : 2.2f ]
+                [ version : 3.4f ],
+                [ version : 2.2f ]
+            ], [
+                [ id: 3 ],
+                [ id: 4 ]
             ])
         then: "Verify update"
             assert db.table(table).field("version").get(3).toFloat() == 3.4f
@@ -340,6 +345,7 @@ abstract class JDBCTest extends Specification {
                     return jdbc.supportsDate ? d.toDate() : d
             }
         when: "Create table"
+            println "Creating table: $table ..."
             assert db.setSQL(getTableCreate(table)) ?: db.setSQL(getTableCreateMulti(table))
         then: "Insert values"
             assert db.table(table).insert([
@@ -371,6 +377,7 @@ abstract class JDBCTest extends Specification {
         expect: "No tables"
             assert db.tables.empty
         when: "Create table"
+            println "Creating table: $table ..."
             assert db.setSQL(getTableCreateMultiplePK(table))
         then : "Be sure the table is there"
             assert db.tables.size() == 1